refactor: 예약 생성 시 이전 날짜 검증 로직 이동
